+++ spec/sleeping_king_studios/tools/object_tools_spec.rb
@@ -474,6 +474,48 @@ RSpec.describe SleepingKingStudios::Tools::ObjectTools do
     end # it
   end # describe
 
+  describe '#object?' do
+    it { expect(instance).to respond_to(:object?).with(1).argument }
+
+    it { expect(described_class).to respond_to(:object?).with(1).argument }
+
+    describe 'with nil' do
+      it { expect(described_class.object? nil).to be true }
+    end # describe
+
+    describe 'with a basic object' do
+      it { expect(described_class.object? BasicObject.new).to be false }
+    end # describe
+
+    describe 'with an object' do
+      it { expect(described_class.object? Object.new).to be true }
+    end # describe
+
+    describe 'with a string' do
+      it { expect(described_class.object? 'greetings,programs').to be true }
+    end # describe
+
+    describe 'with an integer' do
+      it { expect(described_class.object? 42).to be true }
+    end # describe
+
+    describe 'with an empty array' do
+      it { expect(described_class.object? []).to be true }
+    end # describe
+
+    describe 'with a non-empty array' do
+      it { expect(described_class.object? %w(ichi ni san)).to be true }
+    end # describe
+
+    describe 'with an empty hash' do
+      it { expect(described_class.object?({})).to be true }
+    end # describe
+
+    describe 'with a non-empty hash' do
+      it { expect(described_class.object?({ :greetings => 'programs' })).to be true }
+    end # describe
+  end # describe
+
   describe '#try' do
     let(:object) { Object.new }